use framer motion for mobile bottom sheet
