+++ package.json
@@ -16,6 +16,7 @@
     "dayjs": "^1.11.8",
     "eslint": "8.36.0",
     "eslint-config-next": "13.2.4",
+    "framer-motion": "^10.13.1",
     "next": "13.2.4",
     "next-sitemap": "^4.0.7",
     "react": "18.2.0",
